swiftlint fixes
