+++ SmartAILibrary/Strategy/Enums/GovernorType.swift
@@ -73,6 +73,7 @@ public enum GovernorType: Int, Codable {
         let flavors: [Flavor]
     }
 
+    // swiftlint:disable function_body_length
     private func data() -> GovernorTypeData {
 
         switch self {
