+++ SmartAILibrary/Strategy/GameData/PlayerGovernors.swift
@@ -8,120 +8,6 @@
 
 import Foundation
 
-extension Double {
-    static var min     = -Double.greatestFiniteMagnitude
-    // static var MAX_NEG = -Double.leastNormalMagnitude
-    // static var MIN_POS =  Double.leastNormalMagnitude
-    static var max     =  Double.greatestFiniteMagnitude
-
-    var positiveValue: Double {
-
-        if self > 0.0 {
-            return self
-        }
-
-        return 0.0
-    }
-}
-
-public class Governor: Codable {
-
-    enum CodingKeys: CodingKey {
-
-        case type
-        case location
-        case titles
-    }
-
-    public let type: GovernorType
-
-    var location: HexPoint = HexPoint.invalid // <= city
-    var titles: [GovernorTitleType] = []
-
-    // MARK: constructors
-
-    public init(type: GovernorType) {
-
-        self.type = type
-        self.titles.append(self.type.defaultTitle())
-    }
-
-    public required init(from decoder: Decoder) throws {
-
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-
-        self.type = try container.decode(GovernorType.self, forKey: .type)
-        self.location = try container.decode(HexPoint.self, forKey: .location)
-        self.titles = try container.decode([GovernorTitleType].self, forKey: .titles)
-    }
-
-    public func encode(to encoder: Encoder) throws {
-
-        var container = encoder.container(keyedBy: CodingKeys.self)
-
-        try container.encode(self.type, forKey: .type)
-        try container.encode(self.location, forKey: .location)
-        try container.encode(self.titles, forKey: .titles)
-    }
-
-    public func promote(with title: GovernorTitleType) {
-
-        guard !self.titles.contains(title) else {
-            fatalError("cant promote governor with title he/she already has")
-        }
-
-        self.titles.append(title)
-    }
-
-    public func has(title: GovernorTitleType) -> Bool {
-
-        return self.titles.contains(title)
-    }
-
-    public func assign(to cityRef: AbstractCity?) {
-
-        guard let city = cityRef else {
-            fatalError("cant get city")
-        }
-
-        self.location = city.location
-    }
-
-    public func unassign() {
-
-        self.location = HexPoint.invalid
-    }
-
-    public func isAssigned() -> Bool {
-
-        return self.location != HexPoint.invalid
-    }
-
-    public func possiblePromotions() -> [GovernorTitleType] {
-
-        var promotions: [GovernorTitleType] = []
-
-        for title in self.type.titles() {
-
-            if !self.has(title: title) {
-
-                var hasRequired = true
-                for required in title.requiredOr() {
-                    if !self.has(title: required) {
-                        hasRequired = false
-                    }
-                }
-
-                if hasRequired {
-                    promotions.append(title)
-                }
-            }
-        }
-
-        return promotions
-    }
-}
-
 public protocol AbstractPlayerGovernors: AnyObject, Codable {
 
     var player: AbstractPlayer? { get set }
@@ -423,7 +309,7 @@ class PlayerGovernors: AbstractPlayerGovernors {
             fatalError("cant get player")
         }
 
-        var bestGovernorType: GovernorType? = nil
+        var bestGovernorType: GovernorType?
         var bestValue: Int = Int.min
 
         for governorType in GovernorType.all where !self.governors.contains(where: { $0.type == governorType }) {
